requirements.txt
# Core & Web
requests
aiohttp
python-dotenv
beautifulsoup4

# Data Analysis & Numerics
numpy
pandas
scipy
yfinance

# Technical Analysis
TA-Lib
pandas_ta

# Machine Learning & AI
scikit-learn
xgboost
lightgbm
# prophet # این کتابخانه نصب پیچیده‌ای دارد، در صورت نیاز فعال شود
# tensorflow-cpu # برای مدل‌های LSTM/GRU

# Database
psycopg2-binary

# Telegram Bot
python-telegram-bot[httpx]

# Utilities
pytz
redis
asyncio-throttle